chore: Add option to show progress by time in main.py
The code changes in this commit add a new variable `show_by_time` to the `main.py` file. This variable is used to control whether the progress of the program should be shown by waiting for user input at each time step or by using a sleep delay. This enhancement improves the user experience by providing flexibility in how the progress is displayed.